add dccp max-min algorithm
add new method for computing the max-min diverse set.
the method is based on difference of convex functions and concave-convex
procedures.
I believe it will scale at best quadratically with the number of samples
(constraints), which is, however, a lot better than exponential scaling
(including in memory) with the number of variables (as with the current
method).

While this algorithm does not provide exact solutions, it does try to
solve the actual problem, different from the current implementation that
adds points one by one.

See https://stackoverflow.com/a/39969076/1069467